move all app files into app folder
